requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
pypdf==3.12.2
python-docx==0.8.11
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
uvicorn==0.23.1
fastapi==0.100.0
python-jose==3.3.0
passlib==1.7.4